Split noble-migration check logic into reusable lib
As part of the upgrade script, we want to run the check one last time to
ensure that everything is ready to go. Instead of shelling out to it,
move the logc into a Rust library that can be shared by both binaries.
